All sorts of little things

+++ scalars/src/test_scalar_type.cpp
@@ -73,7 +73,7 @@ TEST_F(ScalarTypeTests, BasicInfoFloat) {
     ASSERT_EQ(info.device.device_type, ScalarDeviceType::CPU);
     ASSERT_EQ(info.device.device_id, 0);
     ASSERT_EQ(info.basic_info.bits, sizeof(float)*8);
-    ASSERT_EQ(info.basic_info.code, static_cast<std::uint8_t>(ScalarTypeCode::Float));
+    ASSERT_EQ(info.basic_info.code, ScalarTypeCode::Float);
     ASSERT_EQ(info.basic_info.lanes, 1);
 
 }
@@ -89,7 +89,7 @@ TEST_F(ScalarTypeTests, BasicInfoDouble) {
     ASSERT_EQ(info.device.device_type, ScalarDeviceType::CPU);
     ASSERT_EQ(info.device.device_id, 0);
     ASSERT_EQ(info.basic_info.bits, sizeof(double)*8);
-    ASSERT_EQ(info.basic_info.code, static_cast<std::uint8_t>(ScalarTypeCode::Float));
+    ASSERT_EQ(info.basic_info.code, ScalarTypeCode::Float);
     ASSERT_EQ(info.basic_info.lanes, 1);
 }
